test user service

+++ src/main/java/com/account/manager/repository/UserRepository.java
@@ -1,9 +1,11 @@
 package com.account.manager.repository;
 
+import java.util.Optional;
+
 import com.account.manager.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 };
